Update imports for FacultyFilterState and fix dependency cycle error
A dependency cycle was detected due to the FacultyFilterState being imported from the FacultyPage to the FacultyScheduleTable, and the FacultyScheduleTable being imported to the FacultyPage. Instead, we are defining the FacultyFilterState in the FacultyScheduleTable. Also, the import statements for FacultyFilterState were updated across its usage in the app.
