+++ src/client/classes/index.ts
@@ -1,3 +1,2 @@
 export * from './AppMessage';
 export * from './Validation';
-export * from './Filter';
